+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 - `zeus test` now supports running a glob of tests (`zeus test --env testnet my/path/*.sol`)
 - etherscan verification now integrated.
 - Added the "script" phase, which can run a command as part of an upgrade.
+- Support `zUpdate` for modifying environment state from within a deploy.